+++ lib/demo_tape/runner.rb
@@ -73,8 +73,31 @@ module DemoTape
       commands.select(&:meta?)
     end
 
+    def groups
+      @groups ||=
+        commands
+        .select(&:group?)
+        .each_with_object({}) do |command, buffer|
+          buffer[command.args] = command
+        end
+    end
+
     def executable_commands
       commands.reject(&:meta?)
+    end
+
+    def resolve_group_expansions
+      @commands = commands.flat_map do |command|
+        next command unless command.group_invocation?
+
+        group = groups[command.type]
+
+        unless group
+          command.raise_error("Undefined group: #{command.type.inspect}")
+        end
+
+        group.children
+      end
     end
 
     def resolve_includes
@@ -152,6 +175,7 @@ module DemoTape
       end
 
       resolve_screenshot_count
+      resolve_group_expansions
       run_meta_commands
       resolve_output_paths
       setup_tmp_dir
@@ -679,6 +703,10 @@ module DemoTape
 
       content = File.read(command.args)
       run_type(Command.new("Type", content, *command.options))
+    end
+
+    def run_group(*)
+      # do nothing
     end
 
     def run_screenshot(command)